+++ rtos/os/os.h
@@ -25,7 +25,7 @@
 #define FALSE                       (0)
 #define TRUE                        (1)
 #define PARTIAL                     (2)
-#define MAX_WAIT                    (uint32_t)(-1)
+#define MAX_WAIT                    (uint64_t)(-1)
 
 #ifdef NULL
 #undef NULL
